requirements.txt
pymongo
pycryptodome 
requests
python-dotenv
minio
loguru